requirements.txt
pydub
python-telegram-bot
requests
